fix merge conflict with upstream
Create interface and controller for handling management of roles and
permissions. This is the initial implementation of the feature that lets
an admin create new user and assign them appropriate permissions through
roles. The permissions are currently being generated through seeding.

Some other portions of the codebase(primarily ui related) have been changed as well.